refactor: 점포 대시보드 리뷰 평점 합계 > 평균으로 개선 #49

+++ src/main/java/org/example/luckyburger/domain/shop/dto/response/ShopDashboardResponse.java
@@ -6,12 +6,12 @@ import lombok.Builder;
 public record ShopDashboardResponse(
         Integer todayOrderCount,
         Long todayTotalSales,
-        Float averageRating
+        Double averageRating
 ) {
     public static ShopDashboardResponse of(
             Integer todayOrderCount,
             Long todayTotalSales,
-            Float averageRating) {
+            Double averageRating) {
         return ShopDashboardResponse.builder()
                 .todayOrderCount(todayOrderCount)
                 .todayTotalSales(todayTotalSales)
